# backend/tests/openapi_snapshot.json
{
  "components": {
    "schemas": {
      "AggregationLabelCatalogOut": {
        "properties": {
          "history_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "History Error"
          },
          "history_series_count": {
            "title": "History Series Count",
            "type": "integer"
          },
          "history_status": {
            "enum": [
              "ok",
              "unconfigured",
              "error"
            ],
            "title": "History Status",
            "type": "string"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/AggregationLabelOut"
            },
            "title": "Labels",
            "type": "array"
          },
          "lookback_hours": {
            "enum": [
              24,
              168,
              720
            ],
            "title": "Lookback Hours",
            "type": "integer"
          }
        },
        "required": [
          "lookback_hours",
          "history_status",
          "history_error",
          "history_series_count",
          "labels"
        ],
        "title": "AggregationLabelCatalogOut",
        "type": "object"
      },
      "AggregationLabelOut": {
        "properties": {
          "coverage": {
            "title": "Coverage",
            "type": "number"
          },
          "distinct_count": {
            "title": "Distinct Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "present_count": {
            "title": "Present Count",
            "type": "integer"
          },
          "sample_values": {
            "items": {
              "type": "string"
            },
            "title": "Sample Values",
            "type": "array"
          },
          "sources": {
            "items": {
              "enum": [
                "current",
                "history"
              ],
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "coverage",
          "present_count",
          "total_count",
          "distinct_count",
          "sample_values",
          "sources"
        ],
        "title": "AggregationLabelOut",
        "type": "object"
      },
...
      "AggregationRuleOut": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "group_by_labels": {
            "items": {
              "type": "string"
            },
            "title": "Group By Labels",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "matchers": {
            "items": {
              "$ref": "#/components/schemas/AggregationMatcherIn"
            },
            "title": "Matchers",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "priority": {
            "default": 100,
            "title": "Priority",
            "type": "integer"
          },
          "source_scope": {
            "$ref": "#/components/schemas/SourceScope"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "id",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "AggregationRuleOut",
        "type": "object"
      },
      "AggregationRulePreviewOut": {
        "properties": {
          "by_source": {
            "items": {
              "$ref": "#/components/schemas/AggregationPreviewSourceOut"
            },
            "title": "By Source",
            "type": "array"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/AggregationPreviewGroupOut"
            },
            "title": "Groups",
            "type": "array"
          },
          "matcher_alert_count": {
            "title": "Matcher Alert Count",
            "type": "integer"
          },
          "proposed_group_count": {
            "title": "Proposed Group Count",
            "type": "integer"
          },
          "selected_alert_count": {
            "title": "Selected Alert Count",
            "type": "integer"
          }
        },
        "required": [
          "matcher_alert_count",
          "selected_alert_count",
          "proposed_group_count",
          "groups"
        ],
        "title": "AggregationRulePreviewOut",
        "type": "object"
      },
      "AggregationRulePreviewPut": {
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "group_by_labels": {
            "items": {
              "type": "string"
            },
            "title": "Group By Labels",
            "type": "array"
          },
          "matchers": {
            "items": {
              "$ref": "#/components/schemas/AggregationMatcherIn"
            },
            "title": "Matchers",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "priority": {
            "default": 100,
            "title": "Priority",
            "type": "integer"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "source_scope": {
            "$ref": "#/components/schemas/SourceScope"
          }
        },
        "required": [
          "name"
        ],
        "title": "AggregationRulePreviewPut",
        "type": "object"
      },
      "AggregationRulePut": {
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "group_by_labels": {
            "items": {
              "type": "string"
            },
            "title": "Group By Labels",
            "type": "array"
          },
          "matchers": {
            "items": {
              "$ref": "#/components/schemas/AggregationMatcherIn"
            },
            "title": "Matchers",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "priority": {
            "default": 100,
            "title": "Priority",
            "type": "integer"
          },
          "source_scope": {
            "$ref": "#/components/schemas/SourceScope"
          }
        },
        "required": [
          "name"
        ],
        "title": "AggregationRulePut",
        "type": "object"
      },
      "AlertOut": {
        "properties": {
          "alertname": {
            "title": "Alertname",
            "type": "string"
          },
          "annotations": {
            "additionalProperties": true,
            "title": "Annotations",
            "type": "object"
          },
          "cluster": {
            "title": "Cluster",
            "type": "string"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "evidence_completeness": {
            "title": "Evidence Completeness",
            "type": "string"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "first_seen_at": {
            "title": "First Seen At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "labels": {
            "additionalProperties": true,
            "title": "Labels",
            "type": "object"
          },
          "last_seen_at": {
            "title": "Last Seen At",
            "type": "string"
          },
          "missing_since_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Missing Since At"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "source_id": {
            "default": "legacy",
            "title": "Source Id",
            "type": "string"
          },
          "source_state": {
            "title": "Source State",
            "type": "string"
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          }
        },
        "required": [
          "id",
          "fingerprint",
          "alertname",
          "severity",
          "cluster",
          "labels",
          "annotations",
          "starts_at",
          "ends_at",
          "source_state",
          "missing_since_at",
          "origin",
          "evidence_completeness",
          "first_seen_at",
          "last_seen_at"
        ],
        "title": "AlertOut",
        "type": "object"
      },
...
      "FactOut": {
        "properties": {
          "fact_id": {
            "title": "Fact Id",
            "type": "string"
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          }
        },
        "required": [
          "fact_id",
          "statement"
        ],
        "title": "FactOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvestigationOut": {
        "properties": {
          "facts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FactOut"
            },
            "title": "Facts",
            "type": "array"
          },
          "failure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure"
          },
          "hypotheses": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HypothesisOut"
            },
            "title": "Hypotheses",
            "type": "array"
          },
          "model_calls": {
            "default": 0,
            "title": "Model Calls",
            "type": "integer"
          },
          "model_name": {
            "default": "",
            "title": "Model Name",
            "type": "string"
          },
          "possibly_billed": {
            "default": false,
            "title": "Possibly Billed",
            "type": "boolean"
          },
          "prompt_version": {
            "default": "",
            "title": "Prompt Version",
            "type": "string"
          }
        },
        "title": "InvestigationOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Alert Workbench",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/aggregation-labels": {
      "get": {
        "operationId": "get_aggregation_labels_api_aggregation_labels_get",
        "parameters": [
          {
            "in": "query",
            "name": "lookback_hours",
            "required": false,
            "schema": {
              "default": 168,
              "title": "Lookback Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregationLabelCatalogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Aggregation Labels"
      }
    },
    "/api/aggregation-rules": {
      "get": {
        "operationId": "get_aggregation_rules_api_aggregation_rules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AggregationRuleOut"
                  },
                  "title": "Response Get Aggregation Rules Api Aggregation Rules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Aggregation Rules"
      },
      "post": {
        "operationId": "create_rule_api_aggregation_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AggregationRulePut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregationRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rule"
      }
    },
    "/api/aggregation-rules/preview": {
      "post": {
        "operationId": "preview_rule_api_aggregation_rules_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AggregationRulePreviewPut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregationRulePreviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Rule"
      }
    },
    "/api/aggregation-rules/{rule_id}": {
      "put": {
        "operationId": "update_rule_api_aggregation_rules__rule_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AggregationRulePut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregationRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Rule"
      }
    },
    "/api/alerts/{alert_id}": {
      "get": {
        "operationId": "get_alert_api_alerts__alert_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Alert"
      }
    },
    "/api/alerts/{alert_id}/investigate": {
      "post": {
        "description": "Read one alert's frozen evidence and produce attributed hypotheses.",
        "operationId": "investigate_alert_api_alerts__alert_id__investigate_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Investigate Alert"
      }
    },
...
  }
}